Bug #26864939 - COMPILE WARNINGS IN INNODB PROJECT ON WINDOWS
RB#17462
Reviewed by Bin and Jimmy.

+++ storage/innobase/clone/clone0clone.cc
@@ -603,7 +603,7 @@ Clone_Task_Manager::get_task(
 	heap = get_heap();
 
 	/* Maximum variable length of descriptor. */
-	alloc_len = m_clone_snapshot->get_max_file_name_length();
+	alloc_len = static_cast<uint>(m_clone_snapshot->get_max_file_name_length());
 
 	/* Check with maximum path name length. */
 	if (alloc_len < FN_REFLEN_SE) {
